+++ temba/api/v2/tests.py
@@ -978,7 +978,7 @@ class EndpointsTest(TembaTest):
 
         # try to create new broadcast with no data at all
         response = self.postJSON(url, None, {})
-        self.assertResponseError(response, "text", "This field is required.")
+        self.assertResponseError(response, "non_field_errors", "Must provide either text or attachments.")
 
         # try to create new broadcast with no recipients
         response = self.postJSON(url, None, {"text": "Hello"})
@@ -1024,7 +1024,6 @@ class EndpointsTest(TembaTest):
                 "ticket": str(ticket.uuid),
             },
         )
-
         broadcast = Broadcast.objects.get(id=response.json()["id"])
         self.assertEqual(
             {
@@ -1055,6 +1054,7 @@ class EndpointsTest(TembaTest):
                 "contacts": [self.joe.uuid, self.frank.uuid],
             },
         )
+        self.assertEqual(201, response.status_code)
 
         broadcast = Broadcast.objects.get(id=response.json()["id"])
         self.assertEqual(
@@ -1068,6 +1068,37 @@ class EndpointsTest(TembaTest):
         )
         self.assertEqual("eng", broadcast.base_language)
         self.assertEqual({self.joe, self.frank}, set(broadcast.contacts.all()))
+
+        # create new broadcast without translations containing only text, no attachments
+        response = self.postJSON(
+            url,
+            None,
+            {
+                "text": "Hello",
+                "contacts": [self.joe.uuid, self.frank.uuid],
+            },
+        )
+        self.assertEqual(201, response.status_code)
+
+        broadcast = Broadcast.objects.get(id=response.json()["id"])
+        self.assertEqual({"eng": {"text": "Hello"}}, broadcast.translations)
+
+        # create new broadcast without translations containing only attachments, no text
+        response = self.postJSON(
+            url,
+            None,
+            {
+                "attachments": ["image:http://example.com/test.jpg", "audio/mp3:http://example.com/test.mp3"],
+                "contacts": [self.joe.uuid, self.frank.uuid],
+            },
+        )
+        self.assertEqual(201, response.status_code)
+
+        broadcast = Broadcast.objects.get(id=response.json()["id"])
+        self.assertEqual(
+            {"eng": {"attachments": ["image:http://example.com/test.jpg", "audio/mp3:http://example.com/test.mp3"]}},
+            broadcast.translations,
+        )
 
         # try sending as a flagged org
         self.org.flag()