+++ temba/channels/types/whatsapp_cloud/views.py
@@ -12,7 +12,7 @@ from django.utils.translation import gettext_lazy as _
 from temba.orgs.views import ModalMixin, OrgObjPermsMixin, OrgPermsMixin
 from temba.utils.fields import InputWidget
 from temba.utils.text import truncate
-from temba.utils.views import ContentMenuMixin
+from temba.utils.views import ContentMenuMixin, SpaMixin
 
 from ...models import Channel
 from ...views import ClaimViewMixin
@@ -223,7 +223,7 @@ class ClearSessionToken(OrgPermsMixin, SmartTemplateView):
         return JsonResponse({})
 
 
-class RequestCode(ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActionView):
+class RequestCode(SpaMixin, ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActionView):
     class Form(forms.Form):
         pass
 
@@ -241,6 +241,9 @@ class RequestCode(ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActi
 
     def get_success_url(self):
         return reverse("channels.types.whatsapp_cloud.verify_code", args=[self.object.uuid])
+
+    def derive_menu_path(self):
+        return f"/settings/channels/{self.get_object().uuid}"
 
     def build_content_menu(self, menu):
         obj = self.get_object()
@@ -285,7 +288,7 @@ class RequestCode(ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActi
                 )
 
 
-class VerifyCode(ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActionView):
+class VerifyCode(SpaMixin, ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActionView):
     class Form(forms.Form):
         code = forms.CharField(
             min_length=6, required=True, help_text=_("The 6-digits number verification code"), widget=InputWidget()
@@ -307,6 +310,9 @@ class VerifyCode(ModalMixin, ContentMenuMixin, OrgObjPermsMixin, SmartModelActio
 
     def get_queryset(self):
         return Channel.objects.filter(is_active=True, org=self.request.org, channel_type="WAC")
+
+    def derive_menu_path(self):
+        return f"/settings/channels/{self.get_object().uuid}"
 
     def execute_action(self):